email_classification_project/requirements.txt

flask
pyyaml
spacy
openai
pdfminer.six
python-docx
pymongo
